api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
openai==1.3.8
anthropic==0.7.8
groq==0.4.1
msal==1.25.0
cryptography==41.0.8
psutil==5.9.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
asgi-lifespan==2.1.0